adding households for calendar and shoppinglist
