+++ cook_all_you_can/lib/main.dart
@@ -3,14 +3,13 @@
 // found in the LICENSE file.
 import 'dart:async';
 
-import 'package:cook_all_you_can/index/overview/overview.dart';
-import 'package:cook_all_you_can/index/overview/shared/settings/theme/theme.dart';
-import 'package:cook_all_you_can/index/overview/shared/settings/settings.dart';
 import 'package:cook_all_you_can/index/login.dart';
+import 'package:cook_all_you_can/index/overview/overview/overview.dart';
+import 'package:cook_all_you_can/index/overview/shared/settings/settings.dart';
+import 'package:cook_all_you_can/index/overview/shared/settings/theme/theme.dart';
 import 'package:cook_all_you_can/index/overview/shared/shared.dart';
-import 'package:cook_all_you_can/index/overview/shoppinglist/shoppinglist.dart';
 import 'package:cook_all_you_can/index/overview/shared/user/user.dart';
-import 'package:english_words/english_words.dart';
+import 'package:cook_all_you_can/index/overview/shoppinglist/shoppinglist.dart';
 import 'package:flutter/material.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
