+++ cook_all_you_can/pubspec.yaml
@@ -57,6 +57,7 @@ dependencies:
   flex_color_picker: ^3.3.0
   flutter_colorpicker: ^1.0.3
   flutter_svg: ^2.0.7
+  flutter_launcher_icons: ^0.13.1
 
 dev_dependencies:
   flutter_test:
@@ -68,9 +69,17 @@ dev_dependencies:
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^2.0.0
+  flutter_launcher_icons: ^0.13.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
+
+flutter_icons:
+  image_path: "assets/images/logo_with_new_color.png"
+  android: true
+  ios: true
+  remove_alpha_ios: true
+  resize_behavior: shrink_only
 
 # The following section is specific to Flutter packages.
 flutter:
